// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/SOECompModelDTOs.ts
    export interface IAccountDTO {
        accountId: number;
        accountDimId: number;
        accountNr: string;
        parentAccountId?: number;
        name: string;
        description: string;
        externalCode: string;
        hierarchyOnly: boolean;
        accountTypeSysTermId: number;
        hierarchyNotOnSchedule: boolean;
        accountDim: IAccountDimDTO;
        accountDimNr: number;
        numberName: string;
        dimNameNumberAndName: string;
        amountStop: number;
        unitStop: boolean;
        unit: string;
        rowTextStop: boolean;
        grossProfitCode: number[];
        attestWorkFlowHeadId?: number;
        state: SoeEntityState;
        isAccrualAccount: boolean;
        accountInternals: IAccountInternalDTO[];
        accountIdWithDelimeter: string;
        isAbstract: boolean;
        hasVirtualParent: boolean;
        virtualParentAccountId?: number;
        parentAccounts: IAccountDTO[];
        parentHierachy: Record<number, string>;
        noOParentHierachys: number;
        hierachyId: string;
        hierachyName: string;
        accountHierarchyUniqueId: string;
    }
// ...
    export interface IAccountInternalDTO {
        accountId: number;
        accountNr: string;
        name: string;
        accountDimId: number;
        accountDimNr: number;
        sysSieDimNr?: number;
        sysSieDimNrOrAccountDimNr: number;
        mandatoryLevel: number;
        useVatDeduction: boolean;
        vatDeduction: number;
        account: IAccountDTO;
    }
// ...
    export interface IAccountDimDTO {
        accountDimId: number;
        actorCompanyId: number;
        sysAccountStdTypeParentId?: number;
        parentAccountDimId?: number;
        sysSieDimNr?: number;
        accountDimNr: number;
        name: string;
        shortName: string;
        minChar?: number;
        maxChar?: number;
        linkedToProject: boolean;
        linkedToShiftType: boolean;
        parentAccountDimName: string;
        created?: Date;
        createdBy: string;
        modified?: Date;
        modifiedBy: string;
        state: SoeEntityState;
        useInSchedulePlanning: boolean;
        excludeinAccountingExport: boolean;
        excludeinSalaryReport: boolean;
        useVatDeduction: boolean;
        mandatoryInOrder: boolean;
        mandatoryInCustomerInvoice: boolean;
        onlyAllowAccountsWithParent: boolean;
        allowAccountsWithoutParent: boolean;
        accounts: IAccountDTO[];
        isStandard: boolean;
        isInternal: boolean;
        level: number;
        isSelected: boolean;
    }